Задание 4. Модуляция
